Refactor
- Moved components to folder
- CSS splitted
- Test removed (for now)
