+++ examples/simple-api/src/components/App/App.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Quote from '../Quote/Quote'
+import './App.css'
+
+function App() {
+
+  return (
+    <div className="App">
+      <header className="App-header">
+        <Quote/>
+      </header>
+    </div>
+  );
+}
+
+export default App;
