+++ examples/simple-api/src/components/Quote/Quote.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import WebFont from 'webfontloader';
-import chuckPic from './chuck_norris.png';
-import './App.css';
+import chuckPic from '../../chuck_norris.png';
+import './Quote.css';
 
 WebFont.load({
   google: {
@@ -28,7 +28,7 @@ function Quote() {
   if(err) {
     return (
       <React.Fragment>
-        <h1 className="App-error">Critical Error❗️</h1>
+        <h1 className="error">Critical Error <span role="img" aria-label="Exclamation Mark">❗️</span></h1>
         <h3>Please Reload</h3>
       </React.Fragment>
     )
@@ -40,12 +40,12 @@ function Quote() {
 
   return (
     <React.Fragment>
-      <img src={chuckPic} className="App-pic" alt="logo" />
+      <img src={chuckPic} className="Quote-pic" alt="logo" />
       <p>
       {quote}
       </p>
       <a
-        className="App-link"
+        className="Quote-link"
         href="https://chucknorris.io"
         target="_blank"
         rel="noopener noreferrer"
@@ -56,15 +56,4 @@ function Quote() {
   )
 }
 
-function App() {
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <Quote/>
-      </header>
-    </div>
-  );
-}
-
-export default App;
+export default Quote
